Corrected positioning of slots
Hi, I've just tried to use your app to "recreate" my warehouse which consist of shelves of different width, heights and depths. I've written proper layout.csv file, but it occured that "upper shelves" which should be right above one another or "neighbour shelves" which should be next to each other have gaps between them or are moved among axes. I've corrected few lines of code and the problem was resolved, so I create this pull request if you would like to implemet it to master branch.
	
Thanks for the app.
Best regards,
ZAAN